fix valgrind error in erts/emulator/drivers/unix/ttsl_drv.c
Running some valgrind memory checking showed the error below:

==18040== Thread 6:
==18040== Source and destination overlap in memcpy(0xf3f3f04, 0xf3f3f08, 52)
==18040==    at 0x4C2CFA0: memcpy@@GLIBC_2.14 (in /usr/lib/valgrind/vgpreload_memcheck-amd64-linux.so)
==18040==    by 0x5CF527: del_chars (ttsl_drv.c:845)
==18040==    by 0x5CED5E: ttysl_from_erlang (ttsl_drv.c:658)
==18040==    by 0x4982E3: erts_write_to_port (io.c:1235)
==18040==    by 0x49A2BD: erts_port_command (io.c:2223)
==18040==    by 0x48C054: do_send (bif.c:1962)
==18040==    by 0x48CB6E: erl_send (bif.c:2162)
==18040==    by 0x566599: process_main (beam_emu.c:1665)
==18040==    by 0x4B1A95: sched_thread_func (erl_process.c:4834)
==18040==    by 0x6075E2: thr_wrapper (ethread.c:106)
==18040==    by 0x5560E99: start_thread (pthread_create.c:308)

This occurred on Linux using R15B01 while the shell was emitting a prompt,
but the same problem is still present in R16B.

Change the memcpy on line 845 of ttsl_drv.c to memmove as valgrind
suggests. After the change, verify with valgrind that the error no longer
occurs.