Fix gen_server_SUITE:call_with_huge_message_queue
This test always fails when gen is native compiled, since the
optmization is not implemented then. The test is now skipped when
running with hipe.

The test does also quite often fail on one of the test hosts which
runs in a virtual machine. The reason is that the clock runs "in
bursts" on this host, which is outside of the control of the erlang VM
and probably due to the fact that the host runs as a virtual
machine. To overcome this problem, the echo message is now sent 10000
times instead of 10 times - i.e. the test run is prolonged in order to
even out the effect of the bursts.