Adapt Failed-AVP setting to RFC 6733
When setting these from an #diameter_packet.errors list, select one
Result-Code or {Result-Code, Failed-AVP}, instead of accumulating all
AVP's from the 2-tuples in the list. This is more in keeping with RFC
6733:

  7.5.  Failed-AVP AVP

     The Failed-AVP AVP (AVP Code 279) is of type Grouped and provides
     debugging information in cases where a request is rejected or not
     fully processed due to erroneous information in a specific AVP.  The
     value of the Result-Code AVP will provide information on the reason
     for the Failed-AVP AVP.  A Diameter answer message SHOULD contain an
     instance of the Failed-AVP AVP that corresponds to the error
     indicated by the Result-Code AVP.  For practical purposes, this
     Failed-AVP would typically refer to the first AVP processing error
     that a Diameter node encounters.

The text of RFC 3588 was less specific, not including the last two
sentences.

Note that an improper AVP Length will result in both 5014 and 5005 being
detected for the same AVP. Without this commit, Failed-AVP would be
populated with two AVP's for the same error.