PER, UPER: Fix record name mismatch between HRL file and decoding code
When a SEQUENCE was defined inline inside extension addition group
like this:

InlinedSeq ::= SEQUENCE {
     ...,
     [[
        s SEQUENCE {
	   a INTEGER,
	   b BOOLEAN
        }
     ]]
}

the decoding code would return the contents of the SEQUENCE in a
record named 'InlinedSeq_ExtAddGroup1_s', while the record definition
in the generated HRL file would be 'InlinedSeq_s'.

Since there is no reason to use the longer record name (no risk for
ambiguity), correct the name in the decoding code.