Allow unicode characters for boot and config in init:make_permanent
When release_handler makes a release permanent, it calls
init:make_permanent/2 in order to set the path to the config and boot
file. The paths are coverterd to binaries, and this commit changes
this conversion from using lists_to_binary/1 to using
unicode:characters_to_binary/3 with the encoding set according to the
file name encoding mode (+fnu/fna/fnl).