asn1ct_constructed_per: Simplify gen_enc_line() and friends
The first clause of gen_enc_line() allows us to pass in [] as
the value for Element; if we modify the only caller that passes
[] to pass an actual expression we can remove the first clause.
Furthermore, since the Pos argument was only used by the first
clause, we can remove the Pos argument.

We can also remove the first clause in gen_enc_component_optional(),
since the code in its body is exactly the same as in the following
clause.