erl_interface: add vprintf style functions
As stated at: http://c-faq.com/varargs/handoff.html it is not possible to write
a function which takes a variable number of arguments and passes them to some other function (which takes a variable number of arguments). You should provide a version of that other function which accepts a va_list pointer.

So this provides:
  * ei_x_format => ei_x_vformat
  * ei_x_format_wo_ver => ei_x_vformat_wo_ver
and refactors function implementation in order to avoid code duplication.