Workaround TR gnu/181328, GCC 4.2.1 20070831 on FreeBSD 9.1
The bug manifests so that initialization of an automatic variable
happens after a pointer to the variable is dereferenced, causing
build_exec_return to take the wrong branch by default.

The bug is verified to happen even outside the VM and
is reported to FreeBSD as trouble ID gnu/181328.

The workaround is a slight code rewrite which is optimized differently.

Also removed two warnings about dereferencing type-punned
pointers, which did not affect the bug.