asn1_SUITE: Combine most tests that use External.asn1
By combining most tests that use External.asn,
we will only need to compile that spec once instead of 11 times.

While combining the test cases, make sure to remove the redundant code
in testPrimExternal/1 that tests primitive strings in exactly the same
way as testPrimString/1. Removing the redundant code shaves off a few
seconds of testing time.

Keep testSeqExtension/1 as a separate test case, since it is a
complicated test case that does hairy stuff to the source files.