compiler: Conform returned errors to the documented format
ErrorInfo is documented to be:

  {ErrorLine,Module,ErrorDescriptor}

but for some errors with line numbers it would look like:

  {Module,ErrorDescriptor}

Ensure that all ErrorInfo tuples have three elements. Use 'none'
instead of a line number:

  {none,Module,ErrorDescriptor}

There already are errors that return 'none' when no line number is
available, but that convention was not documented. Mention it in the
documentation.

Also make sure that the compiler will not print 'none' as a line
number in error messages (if the 'report_errors' option is given) as
that looks stupid. That is, when attempting to compile a non-existing
module, the error message should be:

  non-existing.erl: no such file or directory

and not:

  non-existing.erl:none: no such file or directory