asn1_test_lib: Remove unnecessary loading of a compiled ASN.1 spec
asn1ct:compile/2 calls c:c/2 to invoke the compiler and to load
the compiled code. That means that there is no need load the module
after having invoked the ASN.1 compiler.

To make sure that we'll notice if that behavior changes in the future,
extend the rtUI/1 test case to verify that the loaded module has
the correct encoding rule (if asn1ct:compile/2 would fail to reload
the module, the previous encoding rule would be returned).

While at it, it also makes sense to test calling M:bit_string_format/0
to ensure that it continues working.