Remove ^L characters hidden randomly (only) in Erlang code.
Those used in text files as delimiters (RFCs and such) were not removed.
While working on a tool that processes Erlang code and testing it against
this repo, I found out about those little sneaky 0xff. I thought it may be
of help to other people build such tools to remove
non-conforming-to-standard characters.