Simplify and extend diameter_make interface
In particular, make codec/2 flexible as to what's generated, the formats
(erl, hrl, parse, forms and beam) being passed in the options list and
defaulting to [erl, hrl]. The 'parse' format is the internal format to
which dictionaries are parsed, which can be manipulated by flatten/1
before being passed back to codec/2 or format/1.

Remove the (undocumented) dict/1,2 since codec/2 now subsumes it with
the 'parse' option.