Add SASL log quick search for pid in observer
Add a new menu to toggle log view.
Disabled by default.
Disabled if rb_server already started on observed node,
in order not to interfere with somebody else.
If enabled, add a tab in process view where log entries
related to pid process are shown.
Need an observed node with at least a version R16B2,
due to the use of newly capability to rb to write into
a file descriptor (on the observing node).
The file descriptor is a private io server on observer.
Rewrite PID to Observing Node point of view in Log tab.

- Fix some typo in comments.

- CT : Change process to sasl_sup for non empty Log tab.