asn1ct_imm: Add the intermediate instruction {list,List,Dst}
The {list,List,Dst} instruction gives us as general way to
capture the building of something into a variable. That will make
inlining of intermediate code much easier.

It also allows us to eliminate the versions of the apply, call_gen,
and cond instructions that takes a target variable.

Also remove the optimization in the alignment optimization pass
that attempts to replace calls to complete/1 with calls to
iolist_to_binary/1. That optimization will not work anymore without
rewriting, so we will remove it in this commit and introcude a
more powerful optimization in a future commit.